refactor logic into shell.py
