+++ sdk/src/beta9/abstractions/endpoint.py
@@ -1,9 +1,7 @@
 import os
-import socket
 import threading
 import traceback
 import types
-import urllib.parse
 from typing import Any, Callable, List, Optional, Union
 
 from uvicorn.protocols.utils import ClientDisconnected
@@ -16,7 +14,6 @@ from ..abstractions.base.runner import (
     ENDPOINT_DEPLOYMENT_STUB_TYPE,
     ENDPOINT_SERVE_STUB_TYPE,
     ENDPOINT_STUB_TYPE,
-    SHELL_STUB_TYPE,
     RunnerAbstraction,
 )
 from ..abstractions.image import Image
@@ -29,10 +26,8 @@ from ..clients.endpoint import (
     StartEndpointServeResponse,
     StopEndpointServeRequest,
 )
-from ..clients.gateway import GetUrlRequest, GetUrlResponse
-from ..clients.shell import CreateShellRequest, ShellServiceStub
+from ..clients.shell import ShellServiceStub
 from ..env import is_local
-from ..ssh import SSHShell
 from ..type import Autoscaler, GpuType, GpuTypeAlias, QueueDepthAutoscaler, TaskPolicy
 from .mixins import DeployableMixin
 
@@ -611,121 +606,3 @@ class _CallableWrapper(DeployableMixin):
             terminal.error("Serve container failed ❌")
 
         terminal.warn("Endpoint serve timed out. Container has been stopped.")
-
-    @with_grpc_error_handling
-    def shell(self, timeout: int = 0, url_type: str = ""):
-        stub_type = SHELL_STUB_TYPE
-
-        if not self.parent.prepare_runtime(
-            func=self.func, stub_type=stub_type, force_create_stub=True
-        ):
-            return False
-
-        # First, spin up the shell container
-        create_shell_response = self.parent.shell_stub.create_shell(
-            CreateShellRequest(
-                stub_id=self.parent.stub_id,
-                timeout=timeout,
-            )
-        )
-        if not create_shell_response.ok:
-            return terminal.error("Failed to create shell ❌")
-
-        # Then, we can retrieve the URL and issue a CONNECT request / establish a tunnel
-        res: GetUrlResponse = self.parent.gateway_stub.get_url(
-            GetUrlRequest(
-                stub_id=self.parent.stub_id,
-                deployment_id=getattr(self, "deployment_id", ""),
-                url_type=url_type,
-            )
-        )
-        if not res.ok:
-            return terminal.error("Failed to get shell connection URL")
-
-        # Parse the URL to extract the container_id
-        parsed_url = urllib.parse.urlparse(res.url)
-        path_segments = parsed_url.path.split("/")
-        if len(path_segments) < 3:
-            return terminal.error("Invalid URL path")
-
-        container_id = create_shell_response.container_id
-        ssh_token = create_shell_response.token
-
-        # Use the container_id to establish a connection
-        tunnel_socket = None
-        proxy_host, proxy_port = parsed_url.hostname, parsed_url.port
-        try:
-            tunnel_socket = create_connect_tunnel(
-                proxy_host,
-                proxy_port,
-                self.parent.stub_id,
-                container_id,
-                self.parent.config_context.token,
-            )
-        except BaseException:
-            terminal.error(f"Failed to establish ssh tunnel: {traceback.format_exc()}")
-
-        import paramiko
-
-        try:
-            transport = paramiko.Transport(tunnel_socket)
-            transport.start_client()
-            transport.auth_password("runc", ssh_token)
-            session = transport.open_session()
-
-            with SSHShell(
-                channel=session,
-            ) as _:
-                pass
-
-        except paramiko.SSHException as e:
-            print(f"SSH error: {e}")
-        except EOFError:
-            print("Connection closed by the server.")
-        except Exception as e:
-            print(f"An error occurred: {e}")
-        finally:
-            if session:
-                session.close()
-
-            transport.close()
-
-
-def create_connect_tunnel(
-    proxy_host: str, proxy_port: int, stub_id: str, container_id: str, auth_token: str
-) -> socket.socket:
-    """
-    1. Connect to the proxy_host:proxy_port over TCP.
-    2. Send an HTTP CONNECT request for the correct path.
-    3. If we get a 200 response, return the socket as a raw tunnel.
-    """
-
-    s = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-    s.connect((proxy_host, proxy_port))
-
-    # Construct the correct CONNECT request path
-    connect_path = f"/shell/{stub_id}/{container_id}"
-
-    connect_req = (
-        f"CONNECT {connect_path} HTTP/1.1\r\n"
-        f"Host: {proxy_host}:{proxy_port}\r\n"
-        f"Proxy-Connection: Keep-Alive\r\n"
-        f"Authorization: Bearer {auth_token}\r\n"
-        f"\r\n"
-    )
-    s.sendall(connect_req.encode("ascii"))
-
-    response = b""
-    while b"\r\n\r\n" not in response:
-        chunk = s.recv(4096)
-        if not chunk:
-            break
-        response += chunk
-
-    response_str = response.decode("ascii", errors="replace")
-    if "200 OK" not in response_str:
-        s.close()
-        raise Exception(f"CONNECT failed. Response:\n{response_str}")
-
-    # If we reach here, we have a raw TCP tunnel to "container_id"
-    return s
